sgRNA_package/requirements.txt
biopython==1.83
numpy==1.24.4
pandas==2.0.3
tensorflow==2.10.1
joblib==1.4.2
reportlab==3.6.12
xgboost==2.1.4
scikit-learn==1.3.2
